# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_mwork_mwork-api_internal_pkg_response.ErrorInfo:
    properties:
      code:
        type: string
      details:
        additionalProperties:
          type: string
        type: object
      message:
        type: string
    type: object
  github_com_mwork_mwork-api_internal_pkg_response.Meta:
    properties:
      has_next:
        type: boolean
      has_prev:
        type: boolean
      limit:
        type: integer
      page:
        type: integer
      pages:
        type: integer
      total:
        type: integer
    type: object
  github_com_mwork_mwork-api_internal_pkg_response.Response:
    properties:
      data: {}
      error:
        $ref: '#/definitions/github_com_mwork_mwork-api_internal_pkg_response.ErrorInfo'
      meta:
        $ref: '#/definitions/github_com_mwork_mwork-api_internal_pkg_response.Meta'
      success:
        type: boolean
    type: object
...
  internal_domain_profile.EmployerProfileResponse:
    properties:
      castings_posted:
        type: integer
      city:
        type: string
      company_name:
        type: string
      company_type:
        type: string
      contact_person:
        type: string
      contact_phone:
        type: string
      country:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      is_verified:
        type: boolean
      rating:
        type: number
      total_reviews:
        type: integer
      updated_at:
        type: string
      user_id:
        type: string
      website:
        type: string
    type: object
  internal_domain_profile.ModelProfileResponse:
    properties:
      accept_remote_work:
        type: boolean
      age:
        type: integer
      barter_accepted:
        type: boolean
      bio:
        type: string
      categories:
        items:
          type: string
        type: array
      city:
        type: string
      country:
        type: string
      created_at:
        type: string
      experience:
        type: integer
      gender:
        type: string
      height:
        type: number
      hourly_rate:
        type: number
      id:
        type: string
      is_public:
        type: boolean
      languages:
        items:
          type: string
        type: array
      name:
        type: string
      profile_views:
        type: integer
      rating:
        type: number
      skills:
        items:
          type: string
        type: array
      total_reviews:
        type: integer
      travel_cities:
        items:
          type: string
        type: array
      updated_at:
        type: string
      user_id:
        type: string
      visibility:
        type: string
      weight:
        type: number
    type: object
...
  internal_domain_profile.UpdateEmployerProfileRequest:
    properties:
      city:
        maxLength: 100
        minLength: 2
        type: string
      company_name:
        maxLength: 200
        minLength: 2
        type: string
      company_type:
        maxLength: 100
        type: string
      contact_person:
        maxLength: 200
        type: string
      contact_phone:
        maxLength: 20
        type: string
      description:
        maxLength: 2000
        type: string
      website:
        maxLength: 500
        type: string
    type: object
  internal_domain_profile.UpdateModelProfileRequest:
    properties:
      accept_remote_work:
        type: boolean
      age:
        maximum: 100
        minimum: 18
        type: integer
      barter_accepted:
        type: boolean
      bio:
        maxLength: 2000
        type: string
      categories:
        items:
          type: string
        type: array
      city:
        maxLength: 100
        minLength: 2
        type: string
      experience:
        maximum: 50
        minimum: 0
        type: integer
      gender:
        enum:
        - male
        - female
        - other
        type: string
      height:
        maximum: 250
        minimum: 100
        type: number
      hourly_rate:
        minimum: 0
        type: number
      is_public:
        type: boolean
      languages:
        items:
          type: string
        type: array
      name:
        maxLength: 100
        minLength: 2
        type: string
      skills:
        items:
          type: string
        type: array
      travel_cities:
        items:
          type: string
        type: array
      visibility:
        enum:
        - public
        - link_only
        - hidden
        type: string
      weight:
        maximum: 200
        minimum: 30
        type: number
    type: object
...
info:
  contact:
    email: support@swagger.io
    name: API Support
  description: API server for model agency aggregator.
  termsOfService: http://swagger.io/terms/
  title: MWork API
  version: "1.0"
paths:
...
  /profiles/employers/{id}:
    put:
      consumes:
      - application/json
      description: Обновляет профиль работодателя. Доступно только владельцу профиля.
      parameters:
      - description: ID профиля
        in: path
        name: id
        required: true
        type: string
      - description: Обновление профиля работодателя
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_domain_profile.UpdateEmployerProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_mwork_mwork-api_internal_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/internal_domain_profile.EmployerProfileResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_mwork_mwork-api_internal_pkg_response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_mwork_mwork-api_internal_pkg_response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_mwork_mwork-api_internal_pkg_response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_mwork_mwork-api_internal_pkg_response.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/github_com_mwork_mwork-api_internal_pkg_response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_mwork_mwork-api_internal_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: Обновление профиля работодателя
      tags:
      - Profile
  /profiles/me:
    get:
      description: Возвращает профиль текущего пользователя (модель или работодатель).
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_mwork_mwork-api_internal_pkg_response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_mwork_mwork-api_internal_pkg_response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_mwork_mwork-api_internal_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: Мой профиль
      tags:
      - Profile
  /profiles/models:
    get:
      description: Возвращает список профилей моделей с фильтрами и пагинацией.
      parameters:
      - description: Поиск
        in: query
        name: q
        type: string
      - description: Город
        in: query
        name: city
        type: string
      - description: Пол
        in: query
        name: gender
        type: string
      - description: Минимальный возраст
        in: query
        name: age_min
        type: integer
      - description: Максимальный возраст
        in: query
        name: age_max
        type: integer
      - description: Минимальный рост
        in: query
        name: height_min
        type: number
      - description: Максимальный рост
        in: query
        name: height_max
        type: number
      - description: Страница
        in: query
        name: page
        type: integer
      - description: Лимит
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_mwork_mwork-api_internal_pkg_response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/internal_domain_profile.ModelProfileResponse'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_mwork_mwork-api_internal_pkg_response.Response'
      summary: Список моделей
      tags:
      - Profile
  /profiles/models/{id}:
    get:
      description: Возвращает публичный профиль модели по идентификатору.
      parameters:
      - description: ID профиля
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_mwork_mwork-api_internal_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/internal_domain_profile.ModelProfileResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_mwork_mwork-api_internal_pkg_response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_mwork_mwork-api_internal_pkg_response.Response'
      summary: Получение профиля модели
      tags:
      - Profile
    put:
      consumes:
      - application/json
      description: Обновляет профиль модели. Доступно только владельцу профиля.
      parameters:
      - description: ID профиля
        in: path
        name: id
        required: true
        type: string
      - description: Обновление профиля модели
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_domain_profile.UpdateModelProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_mwork_mwork-api_internal_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/internal_domain_profile.ModelProfileResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_mwork_mwork-api_internal_pkg_response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_mwork_mwork-api_internal_pkg_response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_mwork_mwork-api_internal_pkg_response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_mwork_mwork-api_internal_pkg_response.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/github_com_mwork_mwork-api_internal_pkg_response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_mwork_mwork-api_internal_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: Обновление профиля модели
      tags:
      - Profile
  /profiles/models/promoted:
    get:
      description: Возвращает список продвигаемых профилей моделей.
      parameters:
      - description: Город
        in: query
        name: city
        type: string
      - description: Лимит
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_mwork_mwork-api_internal_pkg_response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_mwork_mwork-api_internal_pkg_response.Response'
      summary: Список продвигаемых моделей
      tags:
      - Profile
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"